Add support for Chrome major version > 32
Elements with display set to none are no longer accessible through
getComputedStyle (not being in the DOM). Using CSS visibility will keep
them in the DOM and allow for the breakpoint to work. Note that since
they are in the DOM they will occupy screen space and I assume also
affect the actual height of the container.